Fixing scan host elapsed time

+++ lan_nanny/modules/models/scan_host.py
@@ -22,7 +22,7 @@ class ScanHost(Base):
             },
             {
                 'name': 'elapsed_time',
-                'type': 'int'
+                'type': 'float'
             },
             {
                 'name': 'completed',
